backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.0",
    "langchain-text-splitters>=1.1.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
